% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{brunner}
\alias{brunner}
\title{Identify Atomic Bonds using Brunner's Method}
\usage{
brunner(distances, delta = 1e-04)
}
\arguments{
\item{distances}{A `data.table` of interatomic distances.}

\item{delta}{A small distance offset (default 0.0001).}
}
\value{
A `data.table` of bonded pairs.
}
\description{
(In Progress) An alternative bonding detection method using the
  largest reciprocal gap.
}
\seealso{
Other bonding algorithms: 
\code{\link{hoppe}()},
\code{\link{minimum_distance}()}
}
\concept{bonding algorithms}
