% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Interatomic Distances}
\usage{
calculate_distances(atomic_coordinates, expanded_coords, unit_cell_metrics)
}
\arguments{
\item{atomic_coordinates}{A `data.table` of the primary (asymmetric) atom set.}

\item{expanded_coords}{A `data.table` of atoms in the expanded supercell.}

\item{unit_cell_metrics}{A `data.table` with cell parameters.}
}
\value{
A `data.table` of all non-zero distances.
}
\description{
Computes the distances between a central set of atoms and an
  expanded set, using the metric tensor for accuracy.
}
\examples{
# Example setup
cif_file <- system.file("extdata", "ICSD422.cif", package = "crysmal")
if (file.exists(cif_file)) {
  cif_content <- read_cif_files(cif_file)[[1]]
  atoms <- extract_atomic_coordinates(cif_content)
  metrics <- extract_unit_cell_metrics(cif_content)
  sym_ops <- extract_symmetry_operations(cif_content)
  full_cell <- apply_symmetry_operations(atoms, sym_ops)
  super_cell <- expand_transformed_coords(full_cell)

  dists <- calculate_distances(atoms, super_cell, metrics)
  print(head(dists))
}
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_neighbor_counts}()}
}
\concept{property calculators}
