% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coordinates.R
\name{expand_transformed_coords}
\alias{expand_transformed_coords}
\title{Expand Coordinates into a Supercell}
\usage{
expand_transformed_coords(transformed_coords)
}
\arguments{
\item{transformed_coords}{A `data.table` of atom positions within one unit cell,
typically from `apply_symmetry_operations`.}
}
\value{
A `data.table` containing all atomic positions in the expanded supercell.
}
\description{
Takes a set of atomic coordinates within a single unit cell and
  replicates them into a 3x3x3 supercell grid.
}
\seealso{
Other coordinate processors: 
\code{\link{apply_symmetry_operations}()}
}
\concept{coordinate processors}
