% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{hoppe}
\alias{hoppe}
\title{Identify Atomic Bonds using Hoppe's Method}
\usage{
hoppe(distances, delta = 0.5, tolerance = 0.001)
}
\arguments{
\item{distances}{A `data.table` of interatomic distances.}

\item{delta}{A bond strength cutoff (default 0.5).}

\item{tolerance}{The convergence tolerance for iterative calculation.}
}
\value{
A `data.table` of bonded pairs.
}
\description{
(In Progress) An alternative bonding detection method using
  Effective Coordination Numbers.
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner}()},
\code{\link{minimum_distance}()}
}
\concept{bonding algorithms}
